+++ llm_guard/input_scanners/regex.py
@@ -1,8 +1,6 @@
-from __future__ import annotations
-
 import re
 from enum import Enum
-from typing import Any, Pattern
+from typing import List, Pattern
 
 from presidio_anonymizer.core.text_replace_builder import TextReplaceBuilder
 
@@ -16,9 +14,23 @@ LOGGER = get_logger()
 class MatchType(Enum):
     SEARCH = "search"
     FULL_MATCH = "fullmatch"
+    ALL = "all"
 
-    def match(self, pattern: Pattern[str], text: str) -> Any:
-        return getattr(pattern, self.value)(text)
+    def match(self, pattern: Pattern[str], text: str) -> List[re.Match[str]]:
+        if self.value == "all":
+            return list(pattern.finditer(text))[::-1]  # Reverse order to avoid index issues
+
+        m = None
+        if self.value == "search":
+            m = pattern.search(text)
+
+        if self.value == "fullmatch":
+            m = pattern.fullmatch(text)
+
+        if m is None:
+            return []
+
+        return [m]
 
 
 class Regex(Scanner):
@@ -34,7 +46,7 @@ class Regex(Scanner):
         patterns: list[str],
         *,
         is_blocked: bool = True,
-        match_type: MatchType | str = MatchType.SEARCH,
+        match_type: MatchType | str = MatchType.ALL,
         redact: bool = True,
     ) -> None:
         """
@@ -43,11 +55,11 @@ class Regex(Scanner):
         Parameters:
             patterns (Sequence[str]): A list of regular expressions to use for pattern matching.
             is_blocked (bool): Whether the patterns are blocked or allowed.
-            match_type (str): The type of match to use. Can be either "search" or "fullmatch".
+            match_type (str): The type of match to use.
             redact (bool): Whether to redact the output or not.
 
         Raises:
-            ValueError: If no patterns provided or both good and bad patterns provided.
+            ValueError: If no patterns are provided or both good and bad patterns are provided.
         """
         if isinstance(match_type, str):
             match_type = MatchType(match_type)
@@ -63,19 +75,20 @@ class Regex(Scanner):
     def scan(self, prompt: str) -> tuple[str, bool, float]:
         text_replace_builder = TextReplaceBuilder(original_text=prompt)
         for pattern in self._patterns:
-            match = self._match_type.match(pattern, prompt)
-            if match is None:
+            matches = self._match_type.match(pattern, prompt)
+            if matches is None or len(matches) == 0:
                 continue
 
             if self._is_blocked:
                 LOGGER.warning("Pattern was detected in the text", pattern=pattern)
 
                 if self._redact:
-                    text_replace_builder.replace_text_get_insertion_index(
-                        "[REDACTED]",
-                        match.start(),
-                        match.end(),
-                    )
+                    for match in matches:
+                        text_replace_builder.replace_text_get_insertion_index(
+                            "[REDACTED]",
+                            match.start(),
+                            match.end(),
+                        )
 
                 return text_replace_builder.output_text, False, 1.0