+++ docs/changelog.md
@@ -11,9 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -
 
 ### Fixed
--
+- Regex scanner updates only the first occurrence ([#229](https://github.com/protectai/llm-guard/issues/229)).
 
 ### Changed
+- Upgrade all dependencies.
 - Stop substrings moved to the variables instead of JSON files.
 - **[BREAKING]** New logic to calculate the risk score ([#182](https://github.com/protectai/llm-guard/issues/182)).